Organise code files into appropriate folders
