+++ RefactorThis.Persistence/Entities/Payment.cs
@@ -1,0 +1,8 @@
+namespace RefactorThis.Persistence.Entities
+{
+    public class Payment
+    {
+        public decimal Amount { get; set; }
+        public string Reference { get; set; }
+    }
+}
