+++ RefactorThis.Persistence/Repositories/InvoiceRepository.cs
@@ -1,0 +1,24 @@
+using RefactorThis.Persistence.Entities;
+
+namespace RefactorThis.Persistence.Repositories
+{
+    public class InvoiceRepository
+    {
+        private Invoice _invoice;
+
+        public Invoice GetInvoice(string reference)
+        {
+            return _invoice;
+        }
+
+        public void SaveInvoice(Invoice invoice)
+        {
+            //saves the invoice to the database
+        }
+
+        public void Add(Invoice invoice)
+        {
+            _invoice = invoice;
+        }
+    }
+}
